refactor: collapse the two model resolvers into one
`core/active-model.ts:resolveActiveModelForChat` and
`agent-runtime/resolver.ts:resolveActiveModelRefForChat` walked the
same 5-step chain twice — one returned a string, the other a
`ModelRef`. The ref resolver wrapped the string resolver, called
the catalog a second time to enrich, and was the source of every
\"why are we hitting the catalog twice?\" investigation since Phase
2.1 landed.

Collapsed into one entry point.
`resolveActiveModelForChat(chatId, backend, backendId, config)`
returns `{ model, ref, source }`:

  - `model` — the raw id from the chain (per-chat override →
    backend canonical → operator default → legacy global → null).
  - `ref`   — enriched `ModelRef` for the same id (or `null` when
    `model` is null OR `backendId` isn't a known `BackendId`).
  - `source` — chain-step tag for toast wording.

The enrichment helper (`materialiseRef`) lives in active-model.ts
alongside the chain — one file, one resolver. `getRawModelInfo`
then `resolveModelInfo` fallback ladder is preserved.

`getActiveModelForChat` (string) and `getActiveModelRefForChat`
(ref) are thin convenience wrappers. Callers pick the shape they
need; no double catalog hit either way.

Deletions:
  - `src/core/agent-runtime/resolver.ts` (260 LOC)
  - `src/__tests__/agent-runtime-resolver.test.ts`
  - `agent-runtime/index.ts` exports for the deleted module
  - README section for `resolver.ts` (replaced with the single
    resolver's API doc)

Consumer migration:
  - `frontend/discord/commands.ts`, `frontend/telegram/commands.ts`,
    `frontend/telegram/model-menu.ts` import
    `resolveActiveModelForChat` from `core/active-model.js` and
    destructure `{ ref }` directly. The `{ modelId }` field rename
    folded into `{ model }` since the chain already produces the
    raw id under that name.
  - `active-model.test.ts` switches its exact-match assertions to
    `toMatchObject` so the new `ref` field doesn't break each case.

2807 tests pass, typecheck clean.
