+++ src/__tests__/active-model.test.ts
@@ -138,7 +138,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
       "codex",
       fakeConfig(),
     );
-    expect(result).toEqual({ model: "gpt-5.5", source: "override-valid" });
+    expect(result).toMatchObject({ model: "gpt-5.5", source: "override-valid" });
   });
 
   // ── Step 2 reached: override invalid → backend canonical ───────────
@@ -155,7 +155,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
       "codex",
       fakeConfig(),
     );
-    expect(result).toEqual({
+    expect(result).toMatchObject({
       model: "gpt-5.5",
       source: "override-invalid-fallback",
     });
@@ -210,7 +210,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
       "codex",
       fakeConfig(),
     );
-    expect(result).toEqual({ model: "gpt-5.5", source: "backend-canonical" });
+    expect(result).toMatchObject({ model: "gpt-5.5", source: "backend-canonical" });
   });
 
   // ── Step 3: operator override in config.backendDefaults ────────────
@@ -227,7 +227,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
         },
       }),
     );
-    expect(result).toEqual({
+    expect(result).toMatchObject({
       model: "meta-llama/llama-3.3-70b-instruct:free",
       source: "config-backend-defaults",
     });
@@ -260,7 +260,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
       "claude",
       fakeConfig({ backend: "claude", model: "claude-opus-4-7" }),
     );
-    expect(result).toEqual({
+    expect(result).toMatchObject({
       model: "claude-opus-4-7",
       source: "config-legacy-global",
     });
@@ -277,7 +277,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
     );
     // config.model belongs to the claude backend; openai-agents
     // without backendDefaults must NOT inherit it.
-    expect(result).toEqual({ model: null, source: "none" });
+    expect(result).toMatchObject({ model: null, source: "none" });
   });
 
   // ── Step 5: null when chain exhausted ──────────────────────────────
@@ -290,7 +290,7 @@ describe("resolveActiveModelForChat — 5-step chain", () => {
       "openai-agents",
       fakeConfig({ backend: "claude", model: "" }),
     );
-    expect(result).toEqual({ model: null, source: "none" });
+    expect(result).toMatchObject({ model: null, source: "none" });
   });
 });
